refactor: 상속을 통해 `Dealer`와 `Player`의 중복 코드 제거
